+++ R/contribution.lm.R
@@ -2,11 +2,6 @@
 #' @export
 contribution.lm <-
 function(object, ranker, correction = TRUE, type = "CI", intercept = "exclude") {
-  
-
-    # The ranking variable (wealth, income,...) should be given explicitly.
-    # Throw an error if this is not a numeric one
-    if (!inherits(ranker, "numeric")) stop("Not a numeric ranking variable")
     
     intercept <- match.arg(intercept, choices = c("exclude", "include"))